pyproject.toml
[tool.poetry]
name = "mytest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.8"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-dramatiq = "^0.11.5"
pika = "^1.3.2"
requests = "^2.31.0"
gunicorn = "^21.2.0"
redis = "^5.0.1"
mock = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
